requirements.txt
async-timeout==4.0.2
autopep8==1.7.0
click==8.1.3
cycler==0.11.0
Deprecated==1.2.13
et-xmlfile==1.1.0
flake8==5.0.4
Flask==2.2.2
fonttools==4.36.0
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
mccabe==0.7.0
numpy==1.23.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
six==1.16.0
toml==0.10.2
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
